Bro forgot about negative numbers 💀

+++ src/MMU/RAM.cs
@@ -6,7 +6,7 @@ public class RAM
 {
     private readonly byte[] _memory = new byte[64 * 1024]; // 64KB
 
-    public uint Size => (uint)_memory.Length;
+    public int Size => _memory.Length;
     
     public void Dump(uint offset, uint length)
     {
@@ -31,17 +31,17 @@ public class RAM
         return _memory[address];
     }
     
-    public uint ReadWord(uint address)
+    public int ReadWord(uint address)
     {
         if (address + 4 > _memory.Length)
         {
             throw new IndexOutOfRangeException($"Address out of bounds: {address}");
         }
         
-        return (uint)(_memory[address] |
-                      (_memory[address + 1] << 8) |
-                      (_memory[address + 2] << 16) |
-                      (_memory[address + 3] << 24));
+        return _memory[address] |
+               (_memory[address + 1] << 8) |
+               (_memory[address + 2] << 16) |
+               (_memory[address + 3] << 24);
     }
     
     public void Write(uint address, byte value)
@@ -54,7 +54,7 @@ public class RAM
         _memory[address] = value;
     }
     
-    public void WriteWord(uint address, uint value)
+    public void WriteWord(uint address, int value)
     {
         if (address + 4 > _memory.Length)
         {
@@ -74,7 +74,7 @@ public class RAM
             throw new IndexOutOfRangeException("Program too large");
         }
         
-        Array.Copy(program, 0, _memory, (uint)Segments.TextStart, program.Length);
+        Array.Copy(program, 0, _memory, Segments.TextStart, program.Length);
     }
     
     public void Clear()
